Move hooks to new directory.
