+++ packages/comet-uswds/src/components/modal/index.ts
@@ -1,2 +1,1 @@
 export { default } from './modal';
-export { default as useModal } from './use-modal';
